requirements.txt
Flask==3.0.3
pdfplumber==0.11.4
pypdf2==4.0.1
reportlab==4.2.2
python-dotenv==1.0.1
google-generativeai==0.8.3
